refactor: context 파일에서 hook 분리 [#14]
통일성을 위해 context 파일에서 유효성을 검사하는 hook 분리
